Merge master to develop

+++ src/actions/noteActions.js
@@ -1,8 +1,19 @@
+import _ from 'underscore';
 import { database } from '../data/firebase.js';
-import * as types from '../constants/actionTypes.js';
 
-import { createNewNote, getDeletedTags } from '../common/noteHelpers.js';
+import { 
+    createNewNote, 
+    getDeletedTags, 
+    getNotebookCount, 
+    getTagCount, 
+    filterData 
+} from '../common/noteHelpers.js';
+
 import { uniq } from '../common/helpers.js';
+import { deleteNotebook } from '../actions/notebookActions';
+import { deleteTag } from '../actions/tagActions';
+
+import * as types from '../constants/actionTypes.js';
 
 export function getState() {
     return (dispatch, getState) => {
@@ -10,11 +21,13 @@ export function getState() {
     }
 }
 
-export function getNotes() {
-    return dispatch => {
+export function getNotes(user) {
+    return (dispatch) => {
         dispatch(getNotesRequestedAction());
 
-        return database.ref('/notes').once('value', (snap) => {
+        let notesRef = database.ref('notes');
+
+        notesRef.once('value', (snap) => {
             const notes = snap.val();
             dispatch(getNotesFulfilledAction(notes));
         })
@@ -42,14 +55,16 @@ export function getNotes() {
     }
 }*/
 
-export function addNote(note) {
-    return (dispatch) => {
+export function addNote() {
+    return (dispatch, getState) => {
         dispatch(addNoteRequestedAction());
 
+        const user = getState().userData.user;
+        
         const notesRef = database.ref('notes');
 
         let noteRef = notesRef.push();
-        note = createNewNote(noteRef.key);
+        let note = createNewNote(noteRef.key, user);
 
         noteRef.set(note);
         dispatch(addNoteFulfilledAction(note));
@@ -60,6 +75,8 @@ export function editNote(note, obj = null) {
     return (dispatch, getState) => {
         dispatch(editNoteRequestedAction());
 
+        const user = getState().userData.user;
+
         // refs
         const noteRef = database.ref('notes/' + note.id);
         
@@ -67,19 +84,21 @@ export function editNote(note, obj = null) {
             dispatch(editNoteRejectedAction());
             return;
         }
-        
+
         if (obj) {
             /* If notebook not null and value has changed update the notes notebook only */
-            if (obj.notebook && obj.notebook.name !== note.notebook.name) {
-                const noteNotebookRef = noteRef.child('notebook');
+            if (obj.notebook) {
+                if (!note.notebook || note.notebook.name !== obj.notebook.name) {
+                    const noteNotebookRef = noteRef.child('notebook');
 
-                // If has uid and userId
-                noteNotebookRef.set(obj.notebook)
-                    .then(dispatch(editNoteFulfilledAction(note, obj)))
-                    .catch((error) => {
-                        console.error(error);
-                        dispatch(editNoteRejectedAction());
-                    });
+                    // If has uid and userId
+                    noteNotebookRef.set(obj.notebook)
+                        .then(dispatch(editNoteFulfilledAction(note, obj)))
+                        .catch((error) => {
+                            console.error(error);
+                            dispatch(editNoteRejectedAction());
+                        });
+                }
             }
             
             /* Handle when tags are defined */
@@ -132,9 +151,9 @@ export function editNote(note, obj = null) {
                 }
             }
         } else {
+
             // Update the rest of the note data if not editing notebook
-            database.ref('/notes/' + note.id)
-                .update(note)
+            noteRef.update(note)
                 .then(dispatch(editNoteFulfilledAction(note, {})))
                 .catch((error) => {
                     console.error(error);
@@ -144,38 +163,75 @@ export function editNote(note, obj = null) {
     }
 }
 
-export function deleteNote(id) {
-    return dispatch => {
+export function deleteNote(note) {
+    return (dispatch, getState) => {
         dispatch(deleteNoteRequestedAction());
 
-        database.ref('/notes/' + id)
-            .remove()
-            .then(function() {
-                dispatch(deleteNoteFulfilledAction())
-            })
-            .catch((error) => {
-                console.error(error);
+        const user = getState().userData.user;
+        const notesRef = database.ref('notes');
+        
+        const tagsRef = database.ref('tags');
+
+        /**
+         * These things must be done only for user if user exists and otherwise for things 
+         * where uid does not exist...
+         * So, have to check how many notes there are, and if just one remove it and 
+         * all notebooks and tags since none will be assoicated with any notes.
+         * 
+         * Check the note for notebooks and tags,
+         * if they exist, check if they should be removed from their buckets
+         */
+        if (user) {
+            // cant delete notes that aren't yours
+            if (note.uid !== user.uid) {
                 dispatch(deleteNoteRejectedAction());
-            });
+            } else {
+                notesRef.child(note.id)
+                    .remove()
+                    .then(function() {
+                        dispatch(deleteNoteFulfilledAction())
+                    })
+                    .catch((error) => {
+                        console.error(error);
+                        dispatch(deleteNoteRejectedAction());
+                    });
+            }
+        } else {
+            // can't delete users notes
+            if (note.uid !== undefined) {
+                dispatch(deleteNoteRejectedAction());
+            } else {
+                notesRef.child(note.id)
+                    .remove()
+                    .then(function() {
+                        dispatch(deleteNoteFulfilledAction())
+                    })
+                    .catch((error) => {
+                        console.error(error);
+                        dispatch(deleteNoteRejectedAction());
+                    });
+            }
+        }
     }
 }
 
 export function selectNote(note) {
     return (dispatch, getState) => {
         dispatch(selectNoteRequestedAction());
+
         const currentNotes = getState().noteData.notes;
 
         note = currentNotes.filter(function(n) {
             return n.id === note.id;
         })[0];
 
-        database.ref('/notes/' + note.id + '/isEditing/')
+        database.ref('notes/' + note.id + '/isEditing')
             .set(true)
-            .then(dispatch(selectNoteFulfilledAction(note)))
-            .catch((error) => {
-                console.error(error);
-                dispatch(selectNoteRejectedAction());
-            });
+                .then(dispatch(selectNoteFulfilledAction(note)))
+                .catch((error) => {
+                    console.error(error);
+                    dispatch(selectNoteRejectedAction());
+                });
     }
 }
 
@@ -183,11 +239,14 @@ export function resetSelectedNote() {
     return (dispatch, getState) => {
         dispatch(resetSelectedNoteRequestedAction());
         
-        const notes = getState().noteData.notes;
+        const user = getState().userData.user;        
+        let notes = getState().noteData.notes;
+
+        notes = filterData(user, notes);
 
         notes.forEach((n) => {
             if (n.isEditing) {
-                database.ref('/notes/' + n.id + '/isEditing/')
+                database.ref('notes/' + n.id + '/isEditing')
                     .set(false)
                     .then(dispatch(resetSelectedNoteFulfilledAction(n)))
                     .catch((error) => {
@@ -220,10 +279,6 @@ function getNotesFulfilledAction(notes) {
 function addNoteRequestedAction() {
     return { type: types.AddNoteRequested };
 }
-
-/* function addNoteRejectedAction() {
-    return { type: types.AddNoteRejected };
-} */
 
 function addNoteFulfilledAction(note) {
     return { type: types.AddNoteFulfilled, note };
